Contracts & Audit Trail: Create, Edit, Manage, Sign, View, Audit Trail
VIew Contract right after creating the contract. It shows status as Draft + Signature Panel shows No signature added, also you cannot add the signature at this stage.
Now, when its in active or sent, it shows the signature panel as Add digital signature.
When you click the add digital signature, it opens a signature page where you can actually draw/make your signature and then sign the contract.
Once you sign, it shows the contract being enforced with a digital audit trail.
Also, when you edit a contract that is already signed, it’ll show are you sure, as this will invalidate their signatures, if they proceed and update it, it’ll clear their signatures but show in the audit trail that the contract was updated